Moved mailservice from server code to common backend code so that it can also be used in the algorithm store
