+++ vantage6-backend-common/requirements.txt
@@ -1,4 +1,5 @@
 flask==2.2.5
+flask-mail==0.9.1
 Flask-RESTful==0.3.10
 marshmallow==3.19.0
 marshmallow-sqlalchemy==0.29.0
