+++ vantage6-server/requirements.txt
@@ -4,7 +4,6 @@ flask==2.2.5
 Flask-Cors==4.0.1
 Flask-JWT-Extended==4.4.4
 Flask-Principal==0.4.0
-flask-mail==0.9.1
 flask-marshmallow==0.15.0
 Flask-RESTful==0.3.10
 Flask-SocketIO==5.3.6
